Reorganise tests
Moves some older unit tests into their respective structure as to where
they'd be in the tree of the app folder itself, and puts the custom
trait in the root of the tests folder alongside the Laravel one.

Some of the tests using the custom trait needed the use lines updated so
I took the liberty of sorting them at the same time. Ideal opportunity.
