--- tests/Unit/ExampleTest.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Tests\Unit;
-
-use Tests\TestCase;
-
-class ExampleTest extends TestCase
-{
-    /**
-     * A basic test example.
-     */
-    public function testBasicTest()
-    {
-        $this->assertTrue(true);
-    }
-}
